Replaced block components with single component, changed mux to use x and y instead of a path, replaced relative paths in imports, removed angular material from app module
