+++ src/app/app.routes.ts
@@ -1,9 +1,9 @@
 import { RouterModule, Routes } from '@angular/router';
 
-import { EditorViewComponent } from './components/pages/editor-view/editor-view.component';
-import { MemoryViewComponent } from './components/pages/memory-view/memory-view.component';
-import { SettingsViewComponent } from './components/pages/settings-view/settings-view.component';
-import { SimulatorViewComponent } from './components/pages/simulator-view/simulator-view.component';
+import { EditorViewComponent } from '@components/pages/editor-view/editor-view.component';
+import { MemoryViewComponent } from '@components/pages/memory-view/memory-view.component';
+import { SettingsViewComponent } from '@components/pages/settings-view/settings-view.component';
+import { SimulatorViewComponent } from '@components/pages/simulator-view/simulator-view.component';
 
 const routes: Routes = [
   {
