+++ src/app/components/common/svg/bus/bus.component.spec.ts
