+++ src/app/components/common/svg/mux/mux.component.html
@@ -1,0 +1,3 @@
+<svg:path [attr.d]="svg" fill="#00ffff" stroke="#000000" />
+<svg:text [attr.x]="mux.x" [attr.y]="mux.y" font-family="sans-serif" font-size="14px" text-anchor="middle">MUX</svg:text>
+<svg:text [attr.x]="mux.x" [attr.y]="mux.y + 15" font-family="sans-serif" font-size="14px" text-anchor="middle">{{ mux.inputs }}:{{ mux.outputs }}</svg:text>
