+++ src/app/components/common/svg/mux/mux.component.spec.ts
