+++ src/app/components/simulator/bus-monitor/bus-monitor-item/bus-monitor-item.component.html
