+++ src/app/components/simulator/bus-monitor/bus-monitor-item/bus-monitor-item.component.sass
@@ -1,4 +1,4 @@
-@import '../../../../../../styles'
+@import '../../../../../styles'
 
 div.list-item
   padding: 15px
