+++ src/app/components/simulator/bus-monitor/bus-monitor.component.sass
@@ -1,4 +1,4 @@
-@import '../../../../../styles'
+@import '../../../../styles'
 
 div.tab
   @include overlay
