--- src/app/components/simulator/combinational/combinational.component.html
@@ -1,2 +1,0 @@
-<svg:rect fill="#ffaa33" [attr.x]="x" [attr.y]="y" [attr.width]="width" [attr.height]="height" stroke="#000000" />
-<svg:text [attr.x]="x / 2" [attr.y]="y / 2" font-family="sans-serif" font-size="20px" fill="red">{{ name }}</svg:text>
