--- src/app/components/simulator/controller/controller.component.html
@@ -1,1 +1,0 @@
-<svg:rect fill="#ffff00" [attr.x]="x" [attr.y]="y" [attr.width]="width" [attr.height]="height" stroke="#000000" />
