--- src/app/components/simulator/mux/mux.component.html
@@ -1,1 +1,0 @@
-<svg:path [attr.d]="svg" fill="#00ffff" stroke="#000000" />
