--- src/app/components/simulator/stage-register/stage-register.component.html
@@ -1,1 +1,0 @@
-<svg:rect fill="#00ff33" [attr.x]="x" [attr.y]="y" [attr.width]="width" [attr.height]="height" stroke="#000000" />
