--- src/app/components/simulator/stage/stage.component.html
@@ -1,1 +1,0 @@
-<svg:rect fill="#ff0000" [attr.x]="x" [attr.y]="y" [attr.width]="width" [attr.height]="height" stroke="#000000" />
