+++ src/app/models/simulator/architecture.class.ts
@@ -1,15 +1,12 @@
+import { SVGRect } from '@models/simulator/svg/rect.svg.class';
 import { Bus } from './bus/bus.class';
-import { Combinational } from './combinational/combinational.class';
-import { Controller } from './controller/controller.class';
 import { Mux } from './mux/mux.class';
-import { Register } from './register/register.class';
-import { Stage } from './stage/stage.class';
 
 export class Architecture {
   public bus?: Bus[];
-  public combinational?: Combinational[];
-  public controller?: Controller[];
+  public combinational?: SVGRect[];
+  public controller?: SVGRect[];
   public mux?: Mux[];
-  public register?: Register[];
-  public stage?: Stage[];
+  public register?: SVGRect[];
+  public stage?: SVGRect[];
 }
