+++ src/app/services/simulator/architecture/architecture.service.ts
@@ -1,26 +1,23 @@
 import { Injectable } from '@angular/core';
+import { Architecture } from '@models/simulator/architecture.class';
+import { ARCHITECTURE } from '@models/simulator/architecture.model';
+import { Bus } from '@models/simulator/bus/bus.class';
+import { Mux } from '@models/simulator/mux/mux.class';
+import { SVGRect } from '@models/simulator/svg/rect.svg.class';
+import { RequestService } from '@services/simulator/request/request.service';
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
-import { Architecture } from '../../../models/simulator/architecture.class';
-import { ARCHITECTURE } from '../../../models/simulator/architecture.model';
-import { Bus } from '../../../models/simulator/bus/bus.class';
-import { Combinational } from '../../../models/simulator/combinational/combinational.class';
-import { Controller } from '../../../models/simulator/controller/controller.class';
-import { Mux } from '../../../models/simulator/mux/mux.class';
-import { Register } from '../../../models/simulator/register/register.class';
-import { Stage } from '../../../models/simulator/stage/stage.class';
-import { RequestService } from '../request/request.service';
 
 @Injectable()
 export class ArchitectureService {
 
   private _architecture: Architecture;
   private readonly componentClasses: { class: any, name: string, services: any[] }[] = [
-    { class: Bus,           name: 'bus',              services: [ this.requestService ] },
-    { class: Combinational, name: 'combinational',    services: [ ]                     },
-    { class: Controller,    name: 'controller',       services: [ ]                     },
-    { class: Mux,           name: 'mux',              services: [ ]                     },
-    { class: Stage,         name: 'stage',            services: [ ]                     },
-    { class: Register,      name: 'register',         services: [ ]                     },
+    { class: Bus,     name: 'bus',           services: [ this.requestService ] },
+    { class: SVGRect, name: 'combinational', services: [ ]                     },
+    { class: SVGRect, name: 'controller',    services: [ ]                     },
+    { class: Mux,     name: 'mux',           services: [ ]                     },
+    { class: SVGRect, name: 'stage',         services: [ ]                     },
+    { class: SVGRect, name: 'register',      services: [ ]                     },
   ];
 
   public architecture: BehaviorSubject<Architecture> = new BehaviorSubject<Architecture>(null);
@@ -35,6 +32,9 @@ export class ArchitectureService {
         // If this is the first one, make a place for it in the architecture
         if (!this._architecture[componentClass.name]) {
           this._architecture[componentClass.name] = [ ];
+        }
+        if (componentClass.class.name === 'SVGRect') {
+          instance.type = componentClass.name;
         }
         // Create the class instance using the services and JSON instance
         this._architecture[componentClass.name].push(new componentClass.class(...componentClass.services, instance));
