+++ src/app/services/simulator/websocket/websocket.service.ts
@@ -1,4 +1,3 @@
-
 import { Injectable } from '@angular/core';
 import { Subject } from 'rxjs/Subject';
